Update AppTest.java
in this instance it checks to see if a specific class has a given method it. it checks and prints the results of the points allocated for the test to a pdf. the end user doesnt see what the test entailed only the test case and if it passed for example 

JUnit Test Report
 
Main method exists: PASSED
 
Total Points: 10